+++ dse_do_dashboard/dash_app.py
@@ -12,7 +12,7 @@ import os
 from flask_caching import Cache
 import enum
 import diskcache
-from dash.long_callback import DiskcacheLongCallbackManager
+# from dash.long_callback import DiskcacheLongCallbackManager  # Deprecated, see https://dash.plotly.com/background-callbacks
 
 from dse_do_dashboard.utils.dash_common_utils import ScenarioTableSchema
 
@@ -48,8 +48,10 @@ class DashApp(ABC):
         # Long-running callbacks:
         self.enable_long_running_callbacks = enable_long_running_callbacks
         if self.enable_long_running_callbacks:
-            cache = diskcache.Cache("./cache")
-            self.long_callback_manager = DiskcacheLongCallbackManager(cache)
+            print("WARNING: Long-running callbacks are disabled until refactored with updated Dash API.")
+            # TODO: replace by https://dash.plotly.com/background-callbacks
+            # cache = diskcache.Cache("./cache")
+            self.long_callback_manager = None  #DiskcacheLongCallbackManager(cache), deprecated see https://dash.plotly.com/background-callbacks
         else:
             self.long_callback_manager = None